set up layouts with some more edits extra edits: switch from inMemoryPresestance for the not rememeber me to browserSessionStorage - remove redux test counter - add some info to account navbar menu like username and urlphot - update links
