+++ src/components/global/Navbar.tsx
@@ -18,7 +18,7 @@ const Navbar: React.FC = () => {
         <div className="flex md:order-2 space-x-3 md:space-x-0 gap-2 rtl:space-x-reverse">
           {!authUser ? (
             <>
-              <Link to={'/login'}>
+              <Link to={'/auth'}>
 
               <button
                 type="button"
@@ -29,7 +29,7 @@ const Navbar: React.FC = () => {
               </button>
               </Link>
 
-              <Link to={'/register'}>
+              <Link to={'/auth/register'}>
               <button
                 type="button"
                 className="text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:outline-none focus:ring-primary-300 font-medium rounded-lg text-sm px-4 py-2 text-center dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800"
