+++ src/main.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
 import "./index.css";
 
 // redux
@@ -8,20 +7,28 @@ import { Provider } from "react-redux";
 import { store } from "./state/store.ts";
 
 //react router
-
 import {
   createRoutesFromElements,
   createBrowserRouter,
   Route,
   RouterProvider,
 } from "react-router-dom";
+
+// Layouts
 import RootLyout from "./layouts/RootLyout.tsx";
-import Counter from "./components/Counter.tsx";
+import AuthLayout from "./layouts/AuthLyout.tsx";
+import SettingsLayout from "./layouts/SettingsLayout.tsx";
+
+// Pages
+  // Root
+import App from "./App.tsx";
+  // Auth
 import Login from "./pages/auth/Login.tsx";
 import Register from "./pages/auth/Register.tsx";
 import Reset from "./pages/auth/Reset.tsx";
-import SettingsLayout from "./layouts/SettingsLayout.tsx";
-import Profile from "./pages/profile/Profile.tsx";
+  // Settings
+import Profile from "./pages/settings/Profile.tsx";
+
 
 const router = createBrowserRouter(
   createRoutesFromElements(
@@ -37,23 +44,25 @@ const router = createBrowserRouter(
       >
         <Route index element={<App />} />
         <Route
-          path="/test"
-          element={<Counter />}
-          // loader={contactLoader}
-          // action={contactAction}
-        />
-        <Route path="/login" element={<Login />} />
-        <Route path="/register" element={<Register />} />
-        <Route path="/reset-password" element={<Reset />} />
-      </Route>
-      <Route
-        path="settings"
-        element={<SettingsLayout />}
-        // loader={rootLoader}
-        // action={rootAction}
-        // errorElement={<ErrorPage />}
-      >
-        <Route path="profile" element={<Profile />} />
+          path="auth"
+          element={<AuthLayout />}
+          // loader={rootLoader}
+          // action={rootAction}
+          // errorElement={<ErrorPage />}
+        >
+          <Route index element={<Login />} />
+          <Route path="register" element={<Register />} />
+          <Route path="reset-password" element={<Reset />} />
+        </Route>
+        <Route
+          path="settings"
+          element={<SettingsLayout />}
+          // loader={rootLoader}
+          // action={rootAction}
+          // errorElement={<ErrorPage />}
+        >
+          <Route index element={<Profile />} />
+        </Route>
       </Route>
     </Route>
   )
