--- src/pages/settings/Settings.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Settings: React.FC = () => {
-    return (
-        <div>
-            Settings
-        </div>
-    );
-};
-
-export default Settings;
