+++ src/state/store.ts
@@ -1,10 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
-import counterSlice from "./counter/counterSlice";
 import authSlice from "./auth/authSlice";
 
 export const store = configureStore({
     reducer:{
-        counter:counterSlice,
         auth:authSlice,
     },
 });
